tests/submission/PhilipTaylor/canvas/2d.getcontext.unique.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.getcontext.unique</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.getcontext.html">getcontext</a>.unique</h1>
<p class="desc">getContext('2d') returns the same object</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.unique">context.unique</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("getContext('2d') returns the same object");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext('2d'), canvas.getContext('2d'), "canvas.getContext('2d')", "canvas.getContext('2d')");


});
</script>

<div id="log"></div>
</body></html>
